// app/src/lib/contracts/lb_factory/types.ts
export interface SetLBPairImplementationMsg {
  set_lb_pair_implementation: {
    lb_pair_implementation: {
      id: number;
      code_hash: string;
    };
  };
}

export interface SetLBTokenImplementationMsg {
  set_lb_token_implementation: {
    lb_token_implementation: {
      id: number;
      code_hash: string;
    };
  };
}

export interface CreateLBPairMsg {
  create_lb_pair: {
    token_x: TokenType;
    token_y: TokenType;
    active_id: number;
    bin_step: number;
  }
}

export interface SetLBPairIgnoredMsg {
  set_lb_pair_ignored: {
    token_a: TokenType;
    token_b: TokenType;
    bin_step: number;
    ignored: boolean;
  }
}

export interface SetPresetMsg {
  set_preset: {
    bin_step: number;
    base_factor: number;
    filter_period: number;
    decay_period: number;
    reduction_factor: number;
    variable_fee_control: number;
    protocol_share: number;
    max_volatility_accumulator: number;
    is_open: boolean;
  }
}
// ...
export interface AddQuoteAssetMsg {
  add_quote_asset: {
    asset: TokenType;
  }
}